User-Perspective/Langchain/Prompts/prompts/pyproject.toml
[project]
name = "prompts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=2.0.11",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]
